+++ policy/capabilities_test.go
@@ -25,8 +25,9 @@ var (
 	_ helix.FailoverPolicy  = (*policy.LatencyCircuitBreaker)(nil)
 	_ helix.LatencyRecorder = (*policy.LatencyCircuitBreaker)(nil)
 
-	_ helix.ProbeReporter = (*policy.AdaptiveDualWrite)(nil)
-	_ helix.LatchReporter = (*policy.AdaptiveDualWrite)(nil)
+	_ helix.ProbeReporter        = (*policy.AdaptiveDualWrite)(nil)
+	_ helix.LatchReporter        = (*policy.AdaptiveDualWrite)(nil)
+	_ helix.ProbeLatencyReporter = (*policy.AdaptiveDualWrite)(nil)
 
 	_ helix.EventEmitterSetter = (*policy.AdaptiveDualWrite)(nil)
 	_ helix.EventEmitterSetter = (*policy.CircuitBreaker)(nil)